requirements.txt
FlightRadarAPI==1.2.3
python-telegram-bot==20.1
python-dotenv==1.0.0
uvicorn==0.21.1
starlette==0.26.1
matplotlib==3.7.1
numpy==1.24.2